Added first pass at creating/handling new webserver play request.
This is accessed using the following secure request:
/play?c=<camera_name>

For example:
/play?c=front_of_house_road

This request currently sends the details of all recordings for the
specified camera (with start times of all recordings) to the play.ejs
template. My intention is to use this to render a timeline that can be
used to immediately play a recording from the fullscreen live view that
will be shown by default.

I'm not entirely happy with the current orientation handling where
iPhone and iPad will be concerned but this is a starting poitn.