Tidied up play template and completed updates to load recordings JSON more efficeintly.
This now means you can switch between different cameras in the play
template at the same playback time more efficiently. In summary, when
doing this the play template isn't reloaded. Intead the recordings JSON
for the prev/next camera is loaded into the template and playback
resumes from the same time for the new camera.